basics

+++ basics/arithmeticoperations.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 int main(){
-    int x = 10, y = 3;
+    float x = 10, y = 3;
     cout << x+y << endl;
     cout << x-y << endl;
     cout << x*y << endl;
